revno: 3177 message:   (local vn) - Fixed NPE.
committer: Hieu <hieu.hispvietnam@gmail.com>
branch nick: dhis2
timestamp: Mon 2011-03-28 14:23:41 +0700

+++ local/vn/dhis-web-excel-reporting/src/main/java/org/hisp/dhis/reportexcel/configuration/action/SetReportConfigurationAction.java
@@ -29,6 +29,8 @@ package org.hisp.dhis.reportexcel.configuration.action;
 
 import static org.apache.commons.io.FileUtils.deleteDirectory;
 
+import java.io.File;
+
 import org.hisp.dhis.options.SystemSettingManager;
 import org.hisp.dhis.reportexcel.ReportLocationManager;
 
@@ -77,8 +79,13 @@ public class SetReportConfigurationAction
 
     public String execute()
         throws Exception
-    {        
-        deleteDirectory ( reportLocationManager.getReportExcelTemplateDirectory() );
+    {   
+        File templateDir = reportLocationManager.getReportExcelTemplateDirectory();
+        
+        if ( templateDir != null && templateDir.exists() )
+        {
+            deleteDirectory ( reportLocationManager.getReportExcelTemplateDirectory() );
+        }
         
         systemSettingManager.saveSystemSetting( SystemSettingManager.KEY_REPORT_TEMPLATE_DIRECTORY, templateDirectory );
